+++ lincheck/src/jvm/main/org/jetbrains/kotlinx/lincheck/strategy/managed/eventstructure/EventStructureStrategy.kt
@@ -379,8 +379,8 @@ internal class EventStructureStrategy(
             super.beforeThreadStart(threadDescriptor, startingThread, startingThreadDescriptor)
             runInsideIgnoredSection {
             val newThreadId = threadScheduler.getThreadId(startingThread)
-            //NOTE: Main thread is special cased, since it is forked from the "initial" thread
-            //      which is not tracked in the eventstrcuture. Therefore, we want to skip tracking it
+            // NOTE: The main thread is a special case, since it is forked from the "initial" thread,
+            //       which is not tracked in Eventstrcuture. Therefore, we want to skip tracking it
             if (newThreadId != -1 || newThreadId != eventStructure.mainThreadId) {
                 val currentThreadId = threadScheduler.getCurrentThreadId()
                 eventStructure.addThreadForkEvent(currentThreadId, setOf(newThreadId))
